changed unitTests and added swagger
