+++ src/main/java/com/godel/project/javaMasteryProject/JavaMasteryProjectApplication.java
@@ -1,0 +1,12 @@
+package com.godel.project.javaMasteryProject;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JavaMasteryProjectApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(JavaMasteryProjectApplication.class, args);
+    }
+}
