+++ src/main/java/com/godel/project/javaMasteryProject/enums/Gender.java
@@ -1,0 +1,8 @@
+package com.godel.project.javaMasteryProject.enums;
+
+import lombok.*;
+
+@Getter
+public enum Gender {
+    MALE, FEMALE
+}
