+++ src/main/java/com/godel/project/javaMasteryProject/exceptions/ErrorResponse.java
@@ -1,0 +1,10 @@
+package com.godel.project.javaMasteryProject.exceptions;
+
+import lombok.Data;
+
+@Data
+public class ErrorResponse {
+    private Integer status;
+    private String message;
+    private String error;
+}
