+++ src/main/java/com/godel/project/javaMasteryProject/utils/WebConfig.java
@@ -1,0 +1,15 @@
+package com.godel.project.javaMasteryProject.utils;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addFormatters(FormatterRegistry registry) {
+        registry.addConverter(new LocalDateFormatter());
+    }
+}
+
