+++ src/main/java/com/godel/project/javaMasteryProject/utils/interfaces/IEmployeeConverter.java
@@ -1,0 +1,11 @@
+package com.godel.project.javaMasteryProject.utils.interfaces;
+
+import com.godel.project.javaMasteryProject.controllers.dto.EmployeeDto;
+import com.godel.project.javaMasteryProject.dao.entities.EmployeeEntity;
+
+public interface IEmployeeConverter {
+
+    EmployeeEntity toEntity(EmployeeDto employeeDto);
+
+    EmployeeDto toDto(EmployeeEntity employeeEntity);
+}
