--- src/main/java/com/godel/project/javamasteryproject/JavaMasteryProjectApplication.java
@@ -1,12 +1,0 @@
-package com.godel.project.javamasteryproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class JavaMasteryProjectApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(JavaMasteryProjectApplication.class, args);
-    }
-}
