--- src/main/java/com/godel/project/javamasteryproject/controllers/EmployeeController.java
@@ -1,48 +1,0 @@
-package com.godel.project.javamasteryproject.controllers;
-
-import com.godel.project.javamasteryproject.controllers.dto.EmployeeDto;
-import com.godel.project.javamasteryproject.services.interfaces.IEmployeeService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/employees")
-public class EmployeeController {
-
-    private IEmployeeService employeeService;
-
-    @Autowired
-    public EmployeeController(IEmployeeService employeeService) {
-        this.employeeService = employeeService;
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<EmployeeDto> getById(@PathVariable long id) {
-        return new ResponseEntity<>(employeeService.getById(id), HttpStatus.OK);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<EmployeeDto>> getAll() {
-        return new ResponseEntity<>(employeeService.getAll(), HttpStatus.OK);
-    }
-
-    @PostMapping
-    public ResponseEntity<EmployeeDto> create(@RequestBody EmployeeDto employeeDto) {
-        return new ResponseEntity<>(employeeService.create(employeeDto), HttpStatus.CREATED);
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<EmployeeDto> update(@RequestBody EmployeeDto employeeDto, @PathVariable long id) {
-        return new ResponseEntity<>(employeeService.update(employeeDto, id), HttpStatus.OK);
-    }
-
-    @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable long id) {
-        employeeService.delete(id);
-    }
-}
