--- src/main/java/com/godel/project/javamasteryproject/enums/Gender.java
@@ -1,8 +1,0 @@
-package com.godel.project.javamasteryproject.enums;
-
-import lombok.*;
-
-@Getter
-public enum Gender {
-    MALE, FEMALE
-}
