--- src/main/java/com/godel/project/javamasteryproject/exceptions/ErrorResponse.java
@@ -1,10 +1,0 @@
-package com.godel.project.javamasteryproject.exceptions;
-
-import lombok.Data;
-
-@Data
-public class ErrorResponse {
-    private Integer status;
-    private String message;
-    private String error;
-}
