--- src/main/java/com/godel/project/javamasteryproject/exceptions/GlobalControllerExceptionHandler.java
@@ -1,83 +1,0 @@
-package com.godel.project.javamasteryproject.exceptions;
-
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.sql.SQLException;
-import java.time.DateTimeException;
-import java.util.NoSuchElementException;
-
-@ControllerAdvice
-public class GlobalControllerExceptionHandler {
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> missingEmployeeId(NoSuchElementException ex) {
-        ErrorResponse error = new ErrorResponse();
-        error.setStatus(HttpStatus.BAD_REQUEST.value());
-        error.setError(HttpStatus.BAD_REQUEST.getReasonPhrase());
-        error.setMessage(ex.getMessage());
-        return buildError(error);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> illegalArgs(HttpMessageNotReadableException ex) {
-        ErrorResponse error = new ErrorResponse();
-        error.setStatus(HttpStatus.BAD_REQUEST.value());
-        error.setError(HttpStatus.BAD_REQUEST.getReasonPhrase());
-        error.setMessage(ex.getMessage());
-        return buildError(error);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> dateTimeException(DateTimeException ex) {
-        ErrorResponse error = new ErrorResponse();
-        error.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
-        error.setError(HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase());
-        error.setMessage(ex.getMessage());
-        return buildError(error);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> illegalEnumValue(IllegalArgumentException ex) {
-        ErrorResponse error = new ErrorResponse();
-        error.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
-        error.setError(HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase());
-        error.setMessage(ex.getMessage());
-        return buildError(error);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> missingEmployees(EmptyResultDataAccessException ex) {
-        ErrorResponse error = new ErrorResponse();
-        error.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
-        error.setError(HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase());
-        error.setMessage(ex.getMessage());
-        return buildError(error);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> nullPointerException(NullPointerException ex) {
-        ErrorResponse error = new ErrorResponse();
-        error.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
-        error.setError(HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase());
-        error.setMessage(ex.getMessage());
-        return buildError(error);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> SQLProblems(SQLException ex) {
-        ErrorResponse error = new ErrorResponse();
-        error.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
-        error.setError(HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase());
-        error.setMessage(ex.getMessage());
-        return buildError(error);
-    }
-
-    private ResponseEntity<ErrorResponse> buildError(ErrorResponse error) {
-        return new ResponseEntity<>(error, HttpStatus.valueOf(error.getStatus()));
-    }
-}
