--- src/main/java/com/godel/project/javamasteryproject/utils/WebConfig.java
@@ -1,15 +1,0 @@
-package com.godel.project.javamasteryproject.utils;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.format.FormatterRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addFormatters(FormatterRegistry registry) {
-        registry.addConverter(new LocalDateFormatter());
-    }
-}
-
