--- src/main/java/com/godel/project/javamasteryproject/utils/interfaces/IEmployeeConverter.java
@@ -1,11 +1,0 @@
-package com.godel.project.javamasteryproject.utils.interfaces;
-
-import com.godel.project.javamasteryproject.controllers.dto.EmployeeDto;
-import com.godel.project.javamasteryproject.dao.entities.EmployeeEntity;
-
-public interface IEmployeeConverter {
-
-    EmployeeEntity toEntity(EmployeeDto employeeDto);
-
-    EmployeeDto toDto(EmployeeEntity employeeEntity);
-}
