+++ src/main/resources/application.properties
@@ -9,3 +9,5 @@ spring.jpa.hibernate.ddl-auto=create-drop
 spring.sql.init.mode=always
 spring.sql.init.schema-locations=classpath:scripts/createTable.sql
 spring.sql.init.data-locations=classpath:scripts/fillInTheTable.sql
+
+server.servlet.context-path=/springboot/doc
