+++ src/test/java/com/godel/project/javaMasteryProject/JavaMasteryProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.godel.project.javaMasteryProject;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class JavaMasteryProjectApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
