--- src/test/java/com/godel/project/javamasteryproject/JavaMasteryProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package com.godel.project.javamasteryproject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JavaMasteryProjectApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
